requirements.txt
fastapi==0.103.1
jinja2==3.1.2
openai==0.28.0
python-dotenv==1.0.0
PyYAML==6.0.1
tiktoken==0.5.1
uvicorn==0.23.2
werkzeug==2.3.7
